Refactored ProfessorTeamComponent
