+++ frontend/src/main/frontend/src/components/styles/TeamManager.css
@@ -3,9 +3,6 @@
   max-height: 70vh;
 }
 
-.accordionItem {
-}
-
 .accordionTitle {
   background-color: #82dfe8;
   font-size: 20px;
@@ -16,6 +13,20 @@
 
 .accordionTitle:hover {
   cursor: pointer;
+}
+
+.accordionWrapper {
+  display: flex;
+  justify-content: space-between;
+}
+
+.memberWrapper {
+  display: flex;
+  justify-content: space-between;
+}
+
+.accordionHead {
+  display: flex;
 }
 
 .accordionTeamTitle {
@@ -41,6 +52,7 @@
 
 .teamMember {
   margin-left: 7.5%;
+  padding: 5px;
 }
 
 .buttonPlusTeam {
@@ -50,15 +62,43 @@
 }
 
 .emailInput {
-  border-radius: 50px;
+  border-radius: 30px;
   margin-left: 8px;
 }
 
 .teamA {
   margin-left: 8px;
   text-decoration: underline;
+  transition: ease-in-out 0.5s;
+  border-radius: 10px;
 }
 
 .teamA:hover {
   cursor: pointer;
+  background-color: #6c63ff;
+  color: #fff;
+  border-radius: 10px;
+}
+
+.no-team-wrapper {
+  margin-top: 5%;
+}
+
+.no-team-img-wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.no-team-img {
+  height: 300px;
+  width: 400px;
+}
+
+.no-team-description {
+  color: #6c63ff;
+  font-weight: bold;
+  font-size: 1.5em;
+  margin-top: 30px;
 }
